LPD-93885 Flatten validator package and rename country-code KVP accessors
