+++ modules/apps/account/account-validator-vies/src/main/java/com/liferay/account/validator/vies/internal/VIESAccountEntryValidator.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.account.validator.vies.internal.validator;
+package com.liferay.account.validator.vies.internal;
 
 import com.liferay.account.configuration.AccountEntryValidatorConfiguration;
 import com.liferay.account.constants.AccountConstants;
